% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_filter.R
\name{city_filter}
\alias{city_filter}
\title{City Filter}
\usage{
city_filter(city)
}
\arguments{
\item{city}{A string of the name of the city to be reported on}
}
\value{
A tibble with the data for that city
}
\description{
Function to filter by city
}
