% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_data.R
\docType{data}
\name{weather_data}
\alias{weather_data}
\title{Weather Data from Kenya}
\format{
a tibble containing weather data for 5 cities in Kenya. The 5 cities
are: Mombasa, Lodwar, Kitale, Mandera and Malindi.
}
\source{
\url{https://www1.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt}
}
\usage{
weather_data
}
\description{
Weather data from the GHCN database
}
\keyword{datasets}
