% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_report.R
\name{weather_report}
\alias{weather_report}
\title{Weather Report}
\usage{
weather_report(city)
}
\arguments{
\item{city}{A string of the name of the city to be reported on}
}
\value{
A gg object with 4 plots showcasing the 4 different variables;
precipitation, average temperature, maximum temperature, minimum temperature.
}
\description{
Function to create a weather report for a given city
}
