% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_summary.R
\name{weather_summary}
\alias{weather_summary}
\title{Weather Summary}
\usage{
weather_summary(city)
}
\arguments{
\item{city}{A string of the name of the city to be reported on}
}
\value{
A knitr::kable table showcasing the yearly summaries of 4 different variables;
precipitation, average temperature, maximum temperature, minimum temperature.
}
\description{
Function to create a weather table yearly summary for a given city
}
